Eliminate premature abstraction.
As much as I like the idea of reusable charts, the abstraction of a dot chart
here was more hindrance than help. The issue is that whenever I want to
customize the appearance of the chart—such as overriding the tick format—a chart
abstraction forces me to decide whether that parameter should be exposed as a on
the chart, or if post-selection is required. Generic charts are useful in some
contexts, surely, but why go to great lengths designing a "reusable" chart that
is only used in one place?

The cost of generalization is non-zero. Exposing a zillion parameters is
tedious: Do I expose everything pre-emptively, and have dozens of accessors? Or
do I expose them only as-needed, in which case the tedium is amortized across
all future changes? Post-selection is a leaky abstraction: the more you do it,
the more you assume about how the chart is structured. Furthermore, post-
selection is cumbersome if the chart manages the scales. (Especially if the
chart is designed in an abstract way that can display multiple datasets in
different elements simultaneously!)

Also, charts are necessarily slower if they cannot assume as much about the
current state of the display while doing updates. It doesn’t matter much in this
toy example, but it’s unfortunate that the bulk of the update code is dealing
with parameters (such as the outer width and height of the chart) that never
change in practice. Moreover, premature generalization increases code complexity
with no immediate benefit.